+++ Tests/bootstrap.php
@@ -16,7 +16,6 @@ declare(strict_types=1);
  */
 
 $paths = [
-    \dirname(__DIR__, 4) . '/storage/dev-autoload.php',
     \dirname(__DIR__, 3) . '/autoload.php',
     \dirname(__DIR__) . '/vendor/autoload.php',
 ];