+++ src/Interfaces/EventSubscriberInterface.php
@@ -17,36 +17,13 @@ declare(strict_types=1);
 
 namespace BiuradPHP\Events\Interfaces;
 
+use Symfony\Component\EventDispatcher\EventSubscriberInterface as SymfonyEventSubscriber;
+
 /**
- * An EventSubscriber knows itself what events it is interested in.
- * If an EventSubscriber is added to an EventDispatcherInterface, the manager invokes
- * {@link getSubscribedEvents} and registers the subscriber as a listener for all
- * returned events.
+ * {@inheritdoc}
  *
- * @author Guilherme Blanco <guilhermeblanco@hotmail.com>
- * @author Jonathan Wage <jonwage@gmail.com>
- * @author Roman Borschel <roman@code-factory.org>
- * @author Bernhard Schussek <bschussek@gmail.com>
+ * @author Divine Niiquaye Ibok <divineibok@gmail.com>
  */
-interface EventSubscriberInterface
+interface EventSubscriberInterface extends SymfonyEventSubscriber
 {
-    /**
-     * Returns an array of event names this subscriber wants to listen to.
-     *
-     * The array keys are event names and the value can be:
-     *
-     *  * The method name to call (priority defaults to 0)
-     *  * An array composed of the method name to call and the priority
-     *  * An array of arrays composed of the method names to call and respective
-     *    priorities, or 0 if unset
-     *
-     * For instance:
-     *
-     *  * ['eventName' => 'methodName']
-     *  * ['eventName' => ['methodName', $priority]]
-     *  * ['eventName' => [['methodName1', $priority], ['methodName2']]]
-     *
-     * @return array|mixed The event names to listen to
-     */
-    public static function getSubscribedEvents();
 }