+++ src/TraceableEventDispatcher.php
@@ -18,27 +18,189 @@ declare(strict_types=1);
 namespace BiuradPHP\Events;
 
 use Exception;
+use Psr\EventDispatcher\StoppableEventInterface;
+use SplObjectStorage;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
-class TraceableEventDispatcher extends EventDispatcher
+class TraceableEventDispatcher implements Interfaces\EventDispatcherInterface
 {
+    protected $logger;
+
+    /** @var EventDispatcherInterface */
+    private $dispatcher;
+
+    /** @var array */
+    private $eventsLog = [];
+
+    /** @var SplObjectStorage */
+    private $callStack;
+
+    /** @var array */
+    private $wrappedListeners;
+
+    /** @var array */
+    private $orphanedEvents;
+
+    public function __construct(EventDispatcherInterface $dispatcher, LoggerInterface $logger = null)
+    {
+        $this->dispatcher       = $dispatcher;
+        $this->logger           = $logger;
+        $this->wrappedListeners = [];
+        $this->orphanedEvents   = [];
+    }
+
     /**
-     * Stores information about the events
-     * for display in the profiling.
+     * Proxies all method calls to the original event dispatcher.
      *
-     * @var array
+     * @param string $method    The method name
+     * @param array  $arguments The method arguments
+     *
+     * @return mixed
      */
-    protected static $performanceLog = [];
+    public function __call(string $method, array $arguments)
+    {
+        return $this->dispatcher->{$method}(...$arguments);
+    }
 
     /**
      * {@inheritdoc}
-     * @throws Exception
      */
-    public function dispatch($event, array $payload = [])
+    public function addListener(string $eventName, $listener, int $priority = 0): void
     {
-        $timerStart = \microtime(true);
+        $this->dispatcher->addListener($eventName, $listener, $priority);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function addSubscriber(EventSubscriberInterface $subscriber): void
+    {
+        $this->dispatcher->addSubscriber($subscriber);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function removeListener(string $eventName, $listener)
+    {
+        if (isset($this->wrappedListeners[$eventName])) {
+            foreach ($this->wrappedListeners[$eventName] as $index => $wrappedListener) {
+                if ($wrappedListener->getWrappedListener() === $listener) {
+                    $listener = $wrappedListener;
+                    unset($this->wrappedListeners[$eventName][$index]);
+
+                    break;
+                }
+            }
+        }
+
+        return $this->dispatcher->removeListener($eventName, $listener);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function removeSubscriber(EventSubscriberInterface $subscriber)
+    {
+        return $this->dispatcher->removeSubscriber($subscriber);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getListeners(string $eventName = null)
+    {
+        return $this->dispatcher->getListeners($eventName);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getListenerPriority(string $eventName, $listener)
+    {
+        // we might have wrapped listeners for the event (if called while dispatching)
+        // in that case get the priority by wrapper
+        if (isset($this->wrappedListeners[$eventName])) {
+            foreach ($this->wrappedListeners[$eventName] as $index => $wrappedListener) {
+                if ($wrappedListener->getWrappedListener() === $listener) {
+                    return $this->dispatcher->getListenerPriority($eventName, $wrappedListener);
+                }
+            }
+        }
+
+        return $this->dispatcher->getListenerPriority($eventName, $listener);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function hasListeners(string $eventName = null)
+    {
+        return $this->dispatcher->hasListeners($eventName);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function dispatch(object $event, string $eventName = null): object
+    {
+        $eventName = $eventName ?? \get_class($event);
+
+        if (null === $this->callStack) {
+            $this->callStack = new SplObjectStorage();
+        }
+
+        if (null !== $this->logger && $event instanceof StoppableEventInterface && $event->isPropagationStopped()) {
+            $this->logger->debug(
+                \sprintf('The "%s" event is already stopped. No listeners have been called.', $eventName)
+            );
+        }
+
+        $this->preProcess($eventName);
 
         try {
-            return parent::dispatch($event, $payload);
+            $timerStart = \microtime(true);
+
+            try {
+                $this->dispatcher->dispatch($event, $eventName);
+            } finally {
+                $this->setEventsLog($eventName, $timerStart);
+            }
+        } finally {
+            $this->postProcess($eventName);
+        }
+
+        return $event;
+    }
+
+    /**
+     * @return array
+     */
+    public function getCalledListeners()
+    {
+        if (null === $this->callStack) {
+            return [];
+        }
+
+        $called = [];
+
+        foreach ($this->callStack as $listener) {
+            [$eventName] = $this->callStack->getInfo();
+
+            $called[] = $listener->getInfo($eventName);
+        }
+
+        return $called;
+    }
+
+    /**
+     * @return array
+     */
+    public function getNotCalledListeners()
+    {
+        try {
+            $allListeners = $this->getListeners();
         } catch (Exception $e) {
             if (null !== $this->logger) {
                 $this->logger->info(
@@ -47,11 +209,49 @@ class TraceableEventDispatcher extends EventDispatcher
                 );
             }
 
-            throw $e;
-        } finally {
-            // Enable Profiling
-            static::setPerformanceLogs(\is_object($event) ? \get_class($event) : $event, $timerStart);
+            // unable to retrieve the uncalled listeners
+            return [];
         }
+
+        $calledListeners = [];
+
+        if (null !== $this->callStack) {
+            foreach ($this->callStack as $calledListener) {
+                $calledListeners[] = $calledListener->getWrappedListener();
+            }
+        }
+
+        $notCalled = [];
+
+        foreach ($allListeners as $eventName => $listeners) {
+            foreach ($listeners as $listener) {
+                if (!\in_array($listener, $calledListeners, true)) {
+                    if (!$listener instanceof WrappedListener) {
+                        $listener = new WrappedListener($listener, $this);
+                    }
+                    $notCalled[] = $listener->getInfo($eventName);
+                }
+            }
+        }
+
+        \uasort($notCalled, [$this, 'sortNotCalledListeners']);
+
+        return $notCalled;
+    }
+
+    public function getOrphanedEvents(): array
+    {
+        if (!$this->orphanedEvents) {
+            return [];
+        }
+
+        return \array_merge(...\array_values($this->orphanedEvents));
+    }
+
+    public function reset(): void
+    {
+        $this->callStack          = null;
+        $this->orphanedEvents     = [];
     }
 
     /**
@@ -59,9 +259,9 @@ class TraceableEventDispatcher extends EventDispatcher
      *
      * @return array
      */
-    public static function getPerformanceLogs(): array
+    public function getEventsLogs(): array
     {
-        return static::$performanceLog;
+        return $this->eventsLog;
     }
 
     /**
@@ -70,11 +270,100 @@ class TraceableEventDispatcher extends EventDispatcher
      * @param string       $eventName
      * @param float|string $timerStart
      */
-    public static function setPerformanceLogs(string $eventName, $timerStart): void
+    private function setEventsLog(string $eventName, $timerStart): void
     {
-        static::$performanceLog[] = [
+        $this->eventsLog[] = [
             'event'    => $eventName,
             'duration' => \number_format((\microtime(true) - $timerStart) * 1000, 2) . 'ms',
         ];
+    }
+
+    private function preProcess(string $eventName): void
+    {
+        if (!$this->dispatcher->hasListeners($eventName)) {
+            $this->orphanedEvents[] = $eventName;
+
+            return;
+        }
+
+        foreach ($this->dispatcher->getListeners($eventName) as $listener) {
+            $priority                             = $this->getListenerPriority($eventName, $listener);
+            $wrappedListener                      = new WrappedListener(
+                $listener instanceof WrappedListener ? $listener->getWrappedListener() : $listener,
+                $this
+            );
+            $this->wrappedListeners[$eventName][] = $wrappedListener;
+            $this->dispatcher->removeListener($eventName, $listener);
+            $this->dispatcher->addListener($eventName, $wrappedListener, $priority);
+            $this->callStack->attach($wrappedListener, \compact('eventName'));
+        }
+    }
+
+    private function postProcess(string $eventName): void
+    {
+        unset($this->wrappedListeners[$eventName]);
+        $skipped = false;
+
+        foreach ($this->dispatcher->getListeners($eventName) as $listener) {
+            if (!$listener instanceof WrappedListener) { // #12845: a new listener was added during dispatch.
+                continue;
+            }
+            // Unwrap listener
+            $priority = $this->getListenerPriority($eventName, $listener);
+            $this->dispatcher->removeListener($eventName, $listener);
+            $this->dispatcher->addListener($eventName, $listener->getWrappedListener(), $priority);
+
+            if (null !== $this->logger) {
+                $context = ['event' => $eventName, 'listener' => $listener->getPretty()];
+            }
+
+            if ($listener->wasCalled()) {
+                if (null !== $this->logger) {
+                    $this->logger->debug('Notified event "{event}" to listener "{listener}".', $context);
+                }
+            } else {
+                $this->callStack->detach($listener);
+            }
+
+            if (null !== $this->logger && $skipped) {
+                $this->logger->debug('Listener "{listener}" was not called for event "{event}".', $context);
+            }
+
+            if ($listener->stoppedPropagation()) {
+                if (null !== $this->logger) {
+                    $this->logger->debug(
+                        'Listener "{listener}" stopped propagation of the event "{event}".',
+                        $context
+                    );
+                }
+
+                $skipped = true;
+            }
+        }
+    }
+
+    private function sortNotCalledListeners(array $a, array $b)
+    {
+        if (0 !== $cmp = \strcmp($a['event'], $b['event'])) {
+            return $cmp;
+        }
+
+        if (\is_int($a['priority']) && !\is_int($b['priority'])) {
+            return 1;
+        }
+
+        if (!\is_int($a['priority']) && \is_int($b['priority'])) {
+            return -1;
+        }
+
+        if ($a['priority'] === $b['priority']) {
+            return 0;
+        }
+
+        if ($a['priority'] > $b['priority']) {
+            return -1;
+        }
+
+        return 1;
     }
 }